requirements.txt
PyYAML~=5.4.1
requests~=2.25.1
google-cloud-storage~=1.9.0